execute rake commands under bundler context[1]
also use manageiq rake binary

[1]
"In some cases, running executables without bundle exec may work, if the executable happens to be installed in your system and does not pull in any gems that conflict with your bundle.

However, this is unreliable and is the source of considerable pain. Even if it looks like it works, it may not work in the future or on another machine."

from http://bundler.io/